app/cards/deck.go
/**
Go playground
- Deck main logic
	- initialize new deck
	- deal cards
	- save cards to file
	- retrieve cards from file
	- shuffle decks
**/
package main

import (
	"fmt"
	"io/ioutil"
	"math/rand"
	"os"
	"strings"
	"time"
)

type deck []string

func newDeck() deck {
	cards := deck{}
	cardSuits := []string{"Spades", "Hearts", "Diamonds", "Clubs"}
	cardValues := []string{"Ace", "Two", "Three", "Four", "Five", "Six", "Seven", "Eight", "Nine", "Ten", "Jack", "Queen", "King"}

	for _, cardSuit := range cardSuits {
		for _, cardValue := range cardValues {
			cards = append(cards, cardValue+" of "+cardSuit)
		}
	}

	return cards
}

func deal(d deck, handSize int) (deck, deck) {
	return d[:handSize], d[handSize:]
}

func (d deck) toString() string {
	return strings.Join([]string(d), ",")
}

func (d deck) storeData(filename string) error {
	err := ioutil.WriteFile(filename, []byte(d.toString()), 0644)
	if err != nil {
		fmt.Println("Error:", err)
		os.Exit(1)
	}
	return err
}

func retrieveData(filename string) deck {
	bs, err := ioutil.ReadFile(filename)
	if err != nil {
		fmt.Println("Error:", err)
		os.Exit(1)
	}
	s := strings.Split(string(bs), ",")
	return deck(s)
}

func (d deck) shuffleDeck() {
	s := rand.NewSource(time.Now().UnixNano())
	r := rand.New(s)

	for i := range d {
		nP := r.Intn(len(d) - 1)
		d[i], d[nP] = d[nP], d[i]
	}
}
